breaking up classes in simulation_blocks into own files
